Delete 1.ovpn
Deleted the example certificate for the fake VPN endpoint that we used to collect the traffic of the phones in one place, so that we could capture it there using a network sniffer. While this file has no security impact whatsoever (the traffic is meant to be collected and captured, and the VPN endpoint only exists in the dedicated network between phones and controller server), it raised alarms with some tools that scan Github repositories. So we'll now move on without the example certificate, and users need to create their own certificate before they can start capturing network traffic.